Create forwarding_log_reader.dart
