+++ lib/tizen_device.dart
@@ -27,6 +27,7 @@ import 'package:meta/meta.dart';
 import 'package:path/path.dart';
 import 'package:process/process.dart';
 
+import 'forwarding_log_reader.dart';
 import 'tizen_build_info.dart';
 import 'tizen_builder.dart';
 import 'tizen_sdk.dart';
@@ -357,11 +358,13 @@ class TizenDevice extends Device {
     final bool traceStartup = platformArgs['trace-startup'] as bool ?? false;
     _logger.printTrace('$this startApp');
 
+    final ForwardingLogReader logReader =
+        await getLogReader() as ForwardingLogReader;
     ProtocolDiscovery observatoryDiscovery;
 
     if (debuggingOptions.debuggingEnabled) {
       observatoryDiscovery = ProtocolDiscovery.observatory(
-        await getLogReader(),
+        logReader,
         portForwarder: portForwarder,
         hostPort: debuggingOptions.hostVmServicePort,
         devicePort: debuggingOptions.deviceVmServicePort,
@@ -402,6 +405,7 @@ class TizenDevice extends Device {
         if (debuggingOptions.useTestFonts) '--use-test-fonts',
         if (debuggingOptions.verboseSystemLogs) '--verbose-logging',
       ],
+      '--tizen-logging-port=${logReader.hostPort}',
     ];
 
     // Create a temp file to be consumed by a launching app.
@@ -415,6 +419,9 @@ class TizenDevice extends Device {
       _logger.printError(stdout.trim());
       return LaunchResult.failed();
     }
+
+    // The logging service becomes available right after the app is launched.
+    await logReader.start();
 
     if (!debuggingOptions.debuggingEnabled) {
       return LaunchResult.succeeded();
@@ -484,31 +491,8 @@ class TizenDevice extends Device {
     }
   }
 
-  DateTime get _currentDeviceTime {
-    try {
-      final RunResult result = runSdbSync(usesSecureProtocol
-          ? <String>['shell', '0', 'getdate']
-          : <String>['shell', 'date', '+""%Y-%m-%d %H:%M:%S""']);
-      // Notice that the result isn't normalized with the actual device time
-      // zone. (Because the %z info is missing from the `getdate` result.)
-      // Using the UTC format (appending 'Z' at the end) just prevents the
-      // result from being affected by the host's time zone.
-      return DateTime.parse('${result.stdout.trim()}Z');
-    } on FormatException catch (error) {
-      _logger.printError(error.toString());
-    } on Exception catch (error) {
-      _logger.printError('Failed to get device time: $error');
-    }
-    return DateTime.fromMillisecondsSinceEpoch(0);
-  }
-
-  DateTime _lastClearLogTime;
-
   @override
-  void clearLogs() {
-    // `sdb dlog -c` is not allowed for non-root users.
-    _lastClearLogTime = _currentDeviceTime;
-  }
+  void clearLogs() {}
 
   /// Source: [AndroidDevice.getLogReader] in `android_device.dart`
   @override
@@ -516,11 +500,7 @@ class TizenDevice extends Device {
     TizenTpk app,
     bool includePastLogs = false,
   }) async {
-    return _logReader ??= await TizenDlogReader.createLogReader(
-      this,
-      _processManager,
-      after: includePastLogs ? _lastClearLogTime : _currentDeviceTime,
-    );
+    return _logReader ??= await ForwardingLogReader.createLogReader(this);
   }
 
   @visibleForTesting
@@ -566,139 +546,6 @@ class TizenDevice extends Device {
   Future<void> dispose() async {
     _logReader?.dispose();
     await _portForwarder?.dispose();
-  }
-}
-
-/// A log reader that reads from `sdb dlog`.
-///
-/// Source: [AdbLogReader] in `android_device.dart`
-class TizenDlogReader extends DeviceLogReader {
-  TizenDlogReader._(
-    this.name,
-    this._device,
-    this._sdbProcess,
-    this._after,
-  ) : assert(_after != null) {
-    _linesController = StreamController<String>.broadcast(
-      onListen: _start,
-      onCancel: _stop,
-    );
-  }
-
-  static Future<TizenDlogReader> createLogReader(
-    TizenDevice device,
-    ProcessManager processManager, {
-    @required DateTime after,
-  }) async {
-    // `sdb dlog -m` is not allowed for non-root users.
-    final List<String> args = device.usesSecureProtocol
-        ? <String>['shell', '0', 'showlog_level', 'time']
-        : <String>['dlog', '-v', 'time', 'ConsoleMessage'];
-
-    final Process process = await processManager.start(device.sdbCommand(args));
-
-    return TizenDlogReader._(device.name, device, process, after);
-  }
-
-  final TizenDevice _device;
-  final Process _sdbProcess;
-  final DateTime _after;
-
-  @override
-  final String name;
-
-  StreamController<String> _linesController;
-
-  @override
-  Stream<String> get logLines => _linesController.stream;
-
-  void _start() {
-    const Utf8Decoder decoder = Utf8Decoder(reportErrors: false);
-    _sdbProcess.stdout
-        .transform<String>(decoder)
-        .transform<String>(const LineSplitter())
-        .listen(_onLine);
-    _sdbProcess.stderr
-        .transform<String>(decoder)
-        .transform<String>(const LineSplitter())
-        .listen(_onLine);
-    unawaited(_sdbProcess.exitCode.whenComplete(() {
-      if (_linesController.hasListener) {
-        _linesController.close();
-      }
-    }));
-  }
-
-  // '00-00 00:00:00.000+0000 '
-  final RegExp _timeFormat =
-      RegExp(r'(\d{2}-\d{2}\s\d{2}:\d{2}:\d{2})\.\d{3}[+-]\d{4}\s');
-
-  // 'I/ConsoleMessage(  PID): '
-  final RegExp _logFormat = RegExp(r'[IWEF]\/.+?\(\s*(\d+)\):\s');
-
-  static const List<String> _filteredTexts = <String>[
-    // Issue: https://github.com/flutter-tizen/engine/issues/91
-    'xkbcommon: ERROR:',
-    "couldn't find a Compose file for locale",
-  ];
-
-  bool _acceptedLastLine = true;
-
-  void _onLine(String line) {
-    // This line might be processed after the subscription is closed but before
-    // sdb stops streaming logs.
-    if (_linesController.isClosed) {
-      return;
-    }
-
-    final Match timeMatch = _timeFormat.firstMatch(line);
-    if (timeMatch != null) {
-      // Chop off the time.
-      line = line.replaceFirst(timeMatch.group(0), '');
-
-      final Match logMatch = _logFormat.firstMatch(line);
-      if (logMatch != null) {
-        if (appPid != null && int.parse(logMatch.group(1)) != appPid) {
-          _acceptedLastLine = false;
-          return;
-        } else if (!_device.usesSecureProtocol) {
-          // TODO(swift-kim): Deal with invalid timestamps on TV devices.
-          final DateTime logTime =
-              DateTime.tryParse('${_after.year}-${timeMatch.group(1)}Z');
-          if (logTime != null && logTime.isBefore(_after)) {
-            _acceptedLastLine = false;
-            return;
-          }
-        }
-        if (_filteredTexts.any((String text) => line.contains(text))) {
-          _acceptedLastLine = false;
-          return;
-        }
-        _acceptedLastLine = true;
-        _linesController.add(line);
-      } else {
-        _acceptedLastLine = false;
-      }
-    } else if (line.startsWith('Buffer main is set') ||
-        line.startsWith('ioctl LOGGER') ||
-        line.startsWith('argc = 4, optind = 3') ||
-        line.startsWith('--------- beginning of')) {
-      _acceptedLastLine = false;
-    } else if (_acceptedLastLine) {
-      // If it doesn't match the log pattern at all, then pass it through if we
-      // passed the last matching line through. It might be a multiline message.
-      _linesController.add(line);
-    }
-  }
-
-  void _stop() {
-    _linesController.close();
-    _sdbProcess?.kill();
-  }
-
-  @override
-  void dispose() {
-    _stop();
   }
 }
